Implementation of reading Census 1940s data
Implemented the ability to read in Census 1940s data (EXT1940USCB.dat) and run the
DAS on the persons data.
The schema has been designed to be maximally comparable to the 2010 DDP
DHC-P schema, and reuses the majority of the components in the system.
The output (MDF_PER.txt) is in MDF format.